feat: add cabinet
