+++ src/components/layout.js
@@ -1,4 +1,5 @@
 import Header from './Header'
+
 import SEO from './seo'
 
 export default function Layout({ children }) {
