+++ src/pages/index.js
@@ -1,13 +1,13 @@
-import Login from '../components/Login'
-import UserDashboard from '../components/UserDashboard'
-import { useAuth } from '../context/AuthContext'
+// import Login from '../../components/Login'
+// import UserDashboard from '../../components/UserDashboard'
+// import { useAuth } from '../../context/AuthContext'
 
 export default function HomePage() {
-  const { currentUser } = useAuth()
+//   const { currentUser } = useAuth()
   return (
     <>
-     {!currentUser && <Login />}
-      {currentUser && <UserDashboard/>}
+     {/* {!currentUser && <Login />}
+      {currentUser && <UserDashboard/>} */}
      </>
   );
 }
